requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.8.0

Werkzeug==3.0.3
python-dotenv==1.0.1
google-generativeai>=0.8.0
google-genai
openai>=1.30.0
pytest>=7.0.0
pytest-cov>=4.0.0
httpx>=0.24.0,<0.28.0
websockets>=11,<16
opencv-python-headless
numpy
mediapipe
scikit-learn
supabase>=2.27.2
moviepy==1.0.3
pymediainfo==5.0.3
# pymediainfo imports pkg_resources; setuptools 80+ may omit it on some installs — pin for compatibility
setuptools==69.5.1

boto3>=1.34.0
fastapi>=0.105.0
uvicorn>=0.34.0
gunicorn>=20.1.0